fix(llm-usage): decode money fields as Decimal to preserve precision
Spec section 3.1 mandates Decimal for money to avoid floating-point
error. Kimi, OpenRouter, and OpenAI previously decoded money as Double
then wrapped via Decimal(_:), which preserved the FP error. Switching
the response struct fields directly to Decimal lets JSONDecoder decode
JSON numbers straight into Decimal and removes the lossy Double bridge.

+++ DockCatApp/DockCat/Core/LLMUsage/Providers/KimiUsageProvider.swift
@@ -24,7 +24,7 @@ struct KimiUsageProvider: LLMUsageProvider {
         let state: ProviderUsageSnapshot.State
         switch result {
         case .success(let decoded):
-            let amount = Decimal(decoded.data.availableBalance)
+            let amount = decoded.data.availableBalance
             state = .success(UsageData(
                 balance: Money(amount: amount, currency: "CNY"),
                 totalSpent: nil,
@@ -42,7 +42,7 @@ private struct KimiBalanceResponse: Decodable {
     let data: BalanceData
 
     struct BalanceData: Decodable {
-        let availableBalance: Double
+        let availableBalance: Decimal
 
         enum CodingKeys: String, CodingKey {
             case availableBalance = "available_balance"
